WIB2TPHandler split completed
