+++ src/wib2/tpg/ProcessAVX2FIR.cpp
@@ -1,0 +1,9 @@
+/**
+ * @file ProcessAVX2.cpp
+ *
+ * This is part of the DUNE DAQ , copyright 2020.
+ * Licensing/copyright details are in the COPYING file that you should have
+ * received with this code.
+ */
+
+#include "fdreadoutlibs/wib2/tpg/ProcessAVX2FIR.hpp"
